+++ sdks/go/opspec/interpreter/call/op/params/param/object/validate.go
@@ -1,15 +1,9 @@
 package object
 
 import (
-	"encoding/json"
-	"errors"
-	"fmt"
-	"strings"
-
 	"github.com/opctl/opctl/sdks/go/data/coerce"
+	"github.com/opctl/opctl/sdks/go/internal/jsonschema"
 	"github.com/opctl/opctl/sdks/go/model"
-	"github.com/opctl/opctl/sdks/go/opspec/interpreter/call/op/params/param/formats"
-	"github.com/xeipuuv/gojsonschema"
 )
 
 // Validate validates a value against an object parameter
@@ -18,58 +12,20 @@ func Validate(
 	constraints map[string]interface{},
 ) []error {
 
-	// register custom format checkers
-	gojsonschema.FormatCheckers.Add("docker-image-ref", formats.DockerImageRefFormatChecker{})
-	gojsonschema.FormatCheckers.Add("integer", formats.IntegerFormatChecker{})
-	gojsonschema.FormatCheckers.Add("semver", formats.SemVerFormatChecker{})
-
 	valueAsObject, err := coerce.ToObject(value)
 	if err != nil {
 		return []error{err}
 	}
 
-	unboxedValueAsObject, err := valueAsObject.Unbox()
+	valueAsObjectUnboxed, err := valueAsObject.Unbox()
 	if err != nil {
 		return []error{
 			err,
 		}
 	}
-
-	// guard no constraints
-	if constraints != nil {
-		errs := []error{}
-
-		valueJSONBytes, err := json.Marshal(unboxedValueAsObject)
-		if err != nil {
-			// handle syntax errors specially
-			return append(
-				errs,
-				fmt.Errorf("error validating parameter: %w", err),
-			)
-		}
-
-		result, err := gojsonschema.Validate(
-			gojsonschema.NewGoLoader(constraints),
-			gojsonschema.NewBytesLoader(valueJSONBytes),
-		)
-		if err != nil {
-			// handle syntax errors specially
-			return append(
-				errs,
-				fmt.Errorf("error validating parameter: %w", err),
-			)
-		}
-
-		for _, errString := range result.Errors() {
-			// enum validation errors include `(root) ` prefix we don't want
-			errs = append(
-				errs,
-				errors.New(strings.TrimPrefix(errString.Description(), "(root) ")),
-			)
-		}
-
-		return errs
-	}
-
-	return []error{}
+	
+	return jsonschema.Validate(
+		valueAsObjectUnboxed,
+		constraints,
+	)
 }