+++ sdks/go/opspec/interpreter/stringarray/pkg.go
@@ -1,0 +1,2 @@
+// Package stringarray exposes functionality for interpreting string arrays
+package stringarray