+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file in
 accordance with
 [![keepachangelog 1.0.0](https://img.shields.io/badge/keepachangelog-1.0.0-brightgreen.svg)](http://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased]
+
+### Added
+
+- Allow containers to register multiple names with opctl DNS (i.e. `container.dnsNames`)
+
+### Deprecated
+
+- `container.name`; use `container.dnsNames` (for registering names with opctl DNS) and/or `call.name` (for tracking in the call graph)
+
 ## [0.1.68] - 2025-03-22
 
 ### Fixed